Update README.md
I use a server monitoring solution (Zabbix) and wanted to use signal-cli for notifications. The monitoring solutions can call external scripts in case of problems but obviously uses it's own user to do this. I read about --config in the source and thought, that other people would benefit from this information as well.