Add parameter to specify DBus name
This way multiple instances of signal-cli can run in daemon-mode in
parallel. This was not possible previously as they tried to create the
same DBus.

This is rather my own quick fix and I guess there exists a cleaner way.
(Maybe let the parser parse the busname directly to `SIGNAL_BUSNAME`.)