implement uploadStickerPack for Dbus; fix isRegistered error
Note: isRegistered (and related methods) for Dbus can throw an
InvalidNumberException when the phone number is incorrectly formatted.
Previously this led to uncaught exceptions. They are now handled.
The problem is in SignalServiceAccountManager.java in the
package org.whispersystems.signalservice.api, which ignores the
first character of a proposed phone number and checks that the
rest is a legitimate int64.

updated documentation