Fixed daemon subcommand and App.java logic
Formerly the daemon subcommand failed to create a SignalControl when
a username was specified. (It worked correctly when no username was
specified, implying that all local usernames should be exported.)

Partly this was because of a logic flaw in App.java, whereby daemon
called with a username was incorrectly classified as a LocalCommand
instead of a MultiLocalCommand.

Removed no-longer-needed globals in App.java and Main.java as a result.